move in-memory common stores to separate module
