+++ module/store/common/inmemory/src/main/java/org/example/age/common/service/store/inmemory/InMemoryPendingStoreFactory.java
@@ -1,4 +1,4 @@
-package org.example.age.common.service.store;
+package org.example.age.common.service.store.inmemory;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import java.lang.reflect.Type;
@@ -8,6 +8,8 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 import org.example.age.api.JsonSerializer;
+import org.example.age.common.service.store.PendingStore;
+import org.example.age.common.service.store.PendingStoreFactory;
 
 /** Factory that creates {@link InMemoryPendingStore}'s. */
 @Singleton
@@ -27,14 +29,15 @@ final class InMemoryPendingStoreFactory implements PendingStoreFactory {
     public <V> PendingStore<V> getOrCreate(String name, TypeReference<V> valueTypeRef) {
         InMemoryPendingStore<?> store =
                 stores.computeIfAbsent(name, n -> new InMemoryPendingStore<>(serializer, valueTypeRef));
-        checkValueTypeConsistent(store, valueTypeRef.getType());
+        checkValueTypeConsistent(name, store, valueTypeRef.getType());
         return (PendingStore<V>) store;
     }
 
     /** Checks that the value type is consistent for a single {@link PendingStore}. */
-    private void checkValueTypeConsistent(InMemoryPendingStore<?> store, Type valueType) {
+    private void checkValueTypeConsistent(String name, InMemoryPendingStore<?> store, Type valueType) {
         if (!store.getValueType().equals(valueType)) {
-            String message = String.format("inconsistent value type:\n%s\n%s", store.getValueType(), valueType);
+            String message = String.format(
+                    "inconsistent value type for %s\n  create: %s\n  get: %s", name, store.getValueType(), valueType);
             throw new IllegalArgumentException(message);
         }
     }
