+++ module/store/common/inmemory/src/main/java/org/example/age/common/service/store/inmemory/InMemoryPendingStoreFactoryModule.java
@@ -1,8 +1,9 @@
-package org.example.age.common.service.store;
+package org.example.age.common.service.store.inmemory;
 
 import dagger.Binds;
 import dagger.Module;
 import org.example.age.api.JsonSerializer;
+import org.example.age.common.service.store.PendingStoreFactory;
 
 /**
  * Dagger module that publishes a binding for {@link PendingStoreFactory}, which creates in-memory stores.
